requirements.txt
Flask==2.0.2
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
marshmallow-sqlalchemy==0.26.1
